add 8 lab and 9 lab

+++ src/lab_7/leap_year.java
@@ -6,6 +6,6 @@ public class leap_year {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         System.out.println("введи год");
-        double year = scanner.nextDouble();
+        int year = scanner.nextInt();
         System.out.println((year % 4==0 && year % 100 != 0) || (year % 400 == 0));
 }}
